[hack][fix] `@epfl/epfl-sti-react-library` dependency breakage
`@epfl/epfl-sti-react-library` states a dependency on `react-router-dom`, which AIUI should be a `peerDependency` instead. As a result, it insists on using its own private copy residing in `node_modules/@epfl/epfl-sti-react-library/node_modules`, which results in its `<Link>`s elements (the ones that `<Base>` puts up when `useReactLinks` is true) being incompatible with your app's `<BrowserRouter>` — with an error message that is sybilline at best (“You should not use <Link> outside a <Router>”).

- Fix `@epfl/epfl-sti-react-library`'s `package.json` via patch-package.
  - ... As it turns out, this unfortunately has no effect (at least under `meteor npm`, which is currently of version 6.14.17). Best guess is because `patch-package` comes into play too late. (From the sound of it, the `postinstall` script, in which patch-package lives, probably runs *after* the nested `node_modules` directory gets populated, hmmm)
- Nuke `node_modules/@epfl/epfl-sti-react-library/node_modules/react-router-dom` in the postinstall script for good measure. As per the import resolution rules, this results in `@epfl/epfl-sti-react-library` using the copy we want of `react-router-dom` (the one from the app).