Feat/cadi staging (#23)
* [fix] Random `0` out of nowhere

It was caused because of the fact that `badPasswordTime` is a string. Using `!! ` before calling the value, it's all good.

Thanks to @ponsfrilus

* [feature] cadi-staging instead of test

Cadi-staging is updated everyday. It's better for us.

* [fix] Missing people and duplicate people

Some people weren't findable in the search bar and some were duplicated. It's now fixed by doing a `SELECT DISTINCT` instead of a simple `SELECT`.