Fix/entries without function (#52)
* [fix] If no proper function, no comma

Fixed the comma at the beginning of the line when the accreditation had no proper function

* [feature] Better chevron/triangle

The little triangle or "chevron" next to the accreds and the adminsIT was a bit ugly.

It's now better. And it's a composant.

* [fix] Empty Admins IT

Not displaying the `Admins IT` collapse if there is no Admins IT for this accreditation.