Fix/search as you type (#89)
* [fix] Autocomplete doing with own filter ????

I didn't know that Autocomplete does his own filter, it works well when options are static, which is not our case.
We have to force Autocomplete to not doing his own filter things with `filterOptions`.
Thanks to [that stackoverflow response](https://stackoverflow.com/a/68506363)

* [refactor] api-preprod →api prod for query search

Since the `query` parameter is now also in the production version of api.epfl.ch, no need to use the preprod version.

* [refactor] Added 1sec timer before api call when typing

I think the problem was that wayyy too many requests were sent to the API.
We start sending request since the third character, so it can spam alot.

Now, we wait for 1sec without any new typing before sending a request.
After some testing, this looks better. But is probably not enough.
Maybe some changes from api.epfl.ch are necessary too, this is subject to change.