[fix] Own email address attribute
It wasn't enough to just check the length of the Array returned by the API.
It was necessary to also check if the authorization's attribute `status` was set to `active`.

So, if at least one of the authorization(s) returned by the API has the status `active`, it means that the user has the right to have an email address.