[bugware] for some Brainy-Smurf nonsense from Google
Rids my browser console log of the following lecture:

```
MUI: A component is changing the uncontrolled value state of Autocomplete to be controlled.
Elements should not switch from uncontrolled to controlled (or vice versa).
Decide between using a controlled or uncontrolled Autocomplete element for the lifetime of the component.
The nature of the state is determined during the first render. It's considered controlled if the value is not `undefined`.
More info: https://fb.me/react-controlled-components
```

(where [MUI](https://mui.com/) ultimately
[means Google](https://en.wikipedia.org/wiki/Material_Design).)

Well I ventured to the suggested link, and would you believe it,
[Facebook
says](https://react.dev/learn/sharing-state-between-components#controlled-and-uncontrolled-components)
that these rules are more like guidelines 🤷‍♂️. Fortunately, [some guy
on the Internet](https://stackoverflow.com/a/72136454) figured it out
for me — Presumably, the poser(s) who penned the aforementioned
lecture don't know about the JavaScript `null` value (or for that
matter, how to tell whether a key is missing, rather than `undefined`,
in the props object they get from their caller)? 🤷‍♂️🤷‍♂️